Move implementation into @metamask/smart-accounts-kit-x402 package
