+++ packages/smart-accounts-kit-x402/CHANGELOG.md
@@ -1,0 +1,3 @@
+# Changelog
+
+All notable changes to `@metamask/smart-accounts-kit-x402` will be documented in this file.
