+++ packages/smart-accounts-kit-x402/LICENSE.APACHE2
@@ -1,0 +1,3 @@
+Apache License
+Version 2.0, January 2004
+http://www.apache.org/licenses/
