+++ packages/smart-accounts-kit-x402/LICENSE.MIT0
@@ -1,0 +1,1 @@
+MIT No Attribution License (MIT-0)
