+++ packages/smart-accounts-kit-x402/eslint.config.mjs
@@ -1,0 +1,2 @@
+// eslint-disable-next-line
+export { default } from '../../shared/config/base.eslint.mjs';
