+++ packages/smart-accounts-kit/src/experimental/index.ts
@@ -5,16 +5,3 @@ export {
   type Environment,
   type DelegationStorageConfig,
 } from './delegationStorage';
-export {
-  x402Erc7710Client,
-  type x402DelegationProvider,
-  type x402DelegationPaymentPayload,
-  type x402PaymentRequirements,
-  type x402PaymentPayloadResult,
-  type x402SchemeNetworkClientLike,
-  type x402Erc7710ClientConfig,
-} from './x402Client';
-export {
-  x402Erc7710Server,
-  type x402Erc7710ServerConfig,
-} from './x402Server';
